requirements.txt
Flask==3.0.2
Flask-WTF==1.2.1
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
Flask-Limiter==3.7.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
python-dotenv==1.0.1
PyJWT==2.8.0
pytest==8.1.1